dssd-furniture-backend/authentication/urls.py
from django.urls import path, include
from .views import UserListView, UserListAPIView, UserDetailAPIView, RegisterView, LogoutAPIView, SetNewPasswordAPIView, VerifyEmail, LoginAPIView, PasswordTokenCheckAPI, RequestPasswordResetEmail, UploadViewSet
from rest_framework_simplejwt.views import (
    TokenRefreshView,
)
from rest_framework import routers
from .views import UploadViewSet

router = routers.DefaultRouter()
router.register(r'upload', UploadViewSet, basename="upload")

urlpatterns = [
    path('register/', RegisterView.as_view(), name="register"),
    path('login/', LoginAPIView.as_view(), name="login"),
    path('logout/', LogoutAPIView.as_view(), name="logout"),
    path('email-verify/', VerifyEmail.as_view(), name="email-verify"),
    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('request-reset-email/', RequestPasswordResetEmail.as_view(),
         name="request-reset-email"),
    path('password-reset/<uidb64>/<token>/',
         PasswordTokenCheckAPI.as_view(), name='password-reset-confirm'),
    path('password-reset-complete', SetNewPasswordAPIView.as_view(),
         name='password-reset-complete'),
    path('users', UserListAPIView.as_view(), name="user"),
    path('users/<int:id>/', UserDetailAPIView.as_view(), name="user-update"),
    path('list', UserListView.as_view(), name="user-search"),
    path('users/', include(router.urls)),
]
